+++ src/main/java/com/sisu/sisu/Dao/IProveedorDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.Proveedor;
 
-public interface IProveedorDao extends CrudRepository<Proveedor, Long> {
+public interface IProveedorDao extends CrudRepository<Proveedor, Integer> {
     
 }
