+++ src/main/java/com/sisu/sisu/Dao/IRecetaRemedioDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.RecetaRemedios;
 
-public interface IRecetaRemedioDao extends CrudRepository<RecetaRemedios, Long> {
+public interface IRecetaRemedioDao extends CrudRepository<RecetaRemedios, Integer> {
     
 }
