+++ src/main/java/com/sisu/sisu/Dao/IRemedioLoteDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.RemedioLote;
 
-public interface IRemedioLoteDao extends CrudRepository<RemedioLote, Long> {
+public interface IRemedioLoteDao extends CrudRepository<RemedioLote, Integer> {
     
 }
