+++ src/main/java/com/sisu/sisu/Dao/IRemediosFarmaciaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.RemediosFarmacia;
 
-public interface IRemediosFarmaciaDao extends CrudRepository<RemediosFarmacia, Long> {
+public interface IRemediosFarmaciaDao extends CrudRepository<RemediosFarmacia, Integer> {
     
 }
