+++ src/main/java/com/sisu/sisu/Dao/ITipoProveedorDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.TipoProveedor;
 
-public interface ITipoProveedorDao extends CrudRepository<TipoProveedor, Long>{
+public interface ITipoProveedorDao extends CrudRepository<TipoProveedor, Integer>{
     
 }
