+++ src/main/java/com/sisu/sisu/Dao/RecetaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.Receta;
 
-public interface RecetaDao extends CrudRepository<Receta, Long>{
+public interface RecetaDao extends CrudRepository<Receta, Integer>{
     
 }
