+++ src/main/java/com/sisu/sisu/Service/RecetaService.java
@@ -7,6 +7,6 @@ import com.sisu.sisu.entitys.Receta;
 public interface RecetaService {
     public List<Receta> listarTodo();
     public void registrarReceta(Receta receta);
-    public Receta buscarRecetaId(Long id);
+    public Receta buscarRecetaId(Integer id);
     public void eliminarReceta(Receta receta);
 }
